Some more factorized code
